% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{checkInputTypes}
\alias{checkInputTypes}
\title{检查输入类型}
\usage{
checkInputTypes(data, edit.cols, input.types)
}
\arguments{
\item{data}{数据}

\item{edit.cols}{编辑框}

\item{input.types}{输入类型}
}
\value{
返回值
}
\description{
检查输入类型
}
\examples{
checkInputTypes()
}
