% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtedit_ui.R
\name{dt_btn}
\alias{dt_btn}
\title{定义按纽}
\usage{
dt_btn(name, btn_name = "_add", btn_title = "新增")
}
\arguments{
\item{name}{对象名称}

\item{btn_name}{按续名称}

\item{btn_title}{按纽标题}
}
\value{
返回值
}
\description{
定义按纽
}
\examples{
dt_btn()
}
