% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTedit-package.R
\docType{package}
\name{DTedit-package}
\alias{DTedit}
\alias{DTedit-package}
\title{DTedit: New Editable DataTables for shiny apps}
\description{
This package extends the DT::DataTable to allow users to create, edit, and delete
    rows from the data table. It is a fork of jbryer/DTedit.
}
\author{
\strong{Maintainer}: Daniel \email{suporte@neoage.com.br}

Authors:
\itemize{
  \item Jason Bryer \email{jason@bryer.org}
}

}
\keyword{internal}
