% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtedit.R
\name{dtedit}
\alias{dtedit}
\alias{build_modal_fields}
\title{Edit a DT table}
\usage{
dtedit(
  input,
  output,
  name,
  rv_dat,
  fields,
  values,
  insert,
  dt_opts = list(pageLength = 10L)
)

build_modal_fields(
  dat,
  cols,
  types = NULL,
  args = NULL,
  name_prefix = "dtedit"
)
}
\arguments{
\item{input, output}{Shiny \code{input}/\code{output} objects}

\item{name}{Output name}

\item{rv_dat}{Data (as \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}})}

\item{fields}{Modal fields}

\item{values}{Values for modal fields}

\item{insert}{Callback function for data insert operations}

\item{dt_opts}{Options for \code{DT}}

\item{dat}{Data (passed as \code{data.frame})}

\item{cols}{Relevant columns}

\item{types}{Shiny input types}

\item{args}{(Optional) input field arguments}

\item{name_prefix}{Prefix for input names}
}
\description{
Functions for adding modal-based editing functionality to a DT table in
shiny.
}
