% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTedit-package.R, R/dtedit2-package.R
\docType{package}
\name{dtedit2-package}
\alias{dtedit2}
\alias{dtedit2-package}
\title{dtedit2: New Editable DataTables for shiny apps}
\description{
This package extends the DT::DataTable to allow users to create, edit, and delete
    rows from the data table. It is a fork of jbryer/DTedit.

Editable DataTables for shiny apps
}
\author{
\strong{Maintainer}: Daniel \email{suporte@neoage.com.br}

Authors:
\itemize{
  \item Jason Bryer \email{jason@bryer.org}
}


\strong{Maintainer}: Daniel \email{suporte@neoage.com.br}

Authors:
\itemize{
  \item Jason Bryer \email{jason@bryer.org}
}

}
\keyword{DT}
\keyword{DTedit}
\keyword{DataTable}
\keyword{internal}
\keyword{package}
\keyword{shiny}
