% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtedit2.R
\name{dtedit2}
\alias{dtedit2}
\title{Title}
\usage{
dtedit2(input, output, name, thedata, view.cols = names(thedata),
  view.captions = view.cols, edit.cols = names(thedata),
  edit.label.cols = edit.cols, input.types, input.choices = NULL,
  selectize = TRUE, modal.size = "l", text.width = "100\%",
  textarea.width = "570px", textarea.height = "200px",
  date.width = "100px", numeric.width = "100px",
  select.width = "100\%", defaultPageLength = 6,
  title.delete = "Delete", title.edit = "Edit", title.add = "New",
  label.delete = "删除", label.edit = "修改",
  label.add = "新增", label.copy = "复制", show.delete = TRUE,
  show.update = TRUE, show.insert = TRUE, show.copy = TRUE,
  callback.delete = function(data, row) { },
  callback.update = function(data, olddata, row) { },
  callback.insert = function(data, row) { }, click.time.threshold = 2,
  datatable.options = list(pageLength = defaultPageLength))
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{name}{outputid 对应uiOutput中的id}

\item{thedata}{data部分}

\item{view.cols}{列表标题，默认取数据的列名}

\item{view.captions}{列表标题显示名称,新增功能}

\item{edit.cols}{编辑列,与thedata严格对应}

\item{edit.label.cols}{编码前台显示的名称}

\item{input.types}{控件类型key=value形式,value来源于shiny*Input}

\item{input.choices}{输入控件的选择值,适用于选择框的内容}

\item{selectize}{默认使用即可，效果更好一点}

\item{modal.size}{窗口大小,m=median,l=large,s=small}

\item{text.width}{文本框大小}

\item{textarea.width}{文本区域控件宽}

\item{textarea.height}{文件区域控制高}

\item{date.width}{日期控件宽}

\item{numeric.width}{数值控件宽}

\item{select.width}{选择控件宽}

\item{defaultPageLength}{每页行数}

\item{title.delete}{删除页标题}

\item{title.edit}{编辑页标题}

\item{title.add}{新增页标题}

\item{label.delete}{删除按纽名称}

\item{label.edit}{编辑按纽名称}

\item{label.add}{新增按纽名称}

\item{label.copy}{复制按纽名称}

\item{show.delete}{是否显示删除按纽}

\item{show.update}{是否显示编辑按纽}

\item{show.insert}{是否显示新增按纽}

\item{show.copy}{是否显示复制按纽}

\item{callback.delete}{删除回调函数}

\item{callback.update}{更新回调函数}

\item{callback.insert}{新增回调函数}

\item{click.time.threshold}{点击按纽时间阈值默认2秒}

\item{datatable.options}{datatable额外可传递的选项https://rstudio.github.io/DT/options.html}
}
\value{
返回值
}
\description{
Title
}
\examples{
dtedit2()
}
