% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtedit_demo.R
\name{dtedit_demo}
\alias{dtedit_demo}
\title{Run a shiny app showing how the DTedit function works.}
\usage{
dtedit_demo(...)
}
\arguments{
\item{...}{arguments pass to \code{runApp}

For example, can launch in showcase mode:

\code{DTedit::dtedit_demo(display.mode = "showcase")}

Note that 'showcase' mode shows all
the .R files in the 'shiny_demo' directory, not
just the .R file used for this demonstration!}
}
\description{
Run a shiny app showing how the DTedit function works.
}
