% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtedit_module.R
\name{dtedit_server}
\alias{dtedit_server}
\title{Shiny Server Module version of DTedit}
\usage{
dtedit_server(id, ...)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{...}{other parameters passed to [DTedit::dtedit].}
}
\description{
Shiny Server Module version of DTedit
}
\seealso{
DTedit::dtedit_ui
}
