% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtedit_test.R
\name{dtedit_test}
\alias{dtedit_test}
\title{test application}
\usage{
dtedit_test(appname = "simple", ...)
}
\arguments{
\item{appname}{choose test
simple
simple_modular
reactive
callback
error_test
selectInputReactive
password
datetimeInput
inputEvent
selectizeInput
logical}

\item{...}{extra options passed to shiny::shinyApp}
}
\value{
a shiny app
}
\description{
for testthat/codecov
}
