% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtedit_module.R
\name{dtedit_ui}
\alias{dtedit_ui}
\title{Shiny UI Module version of DTedit}
\usage{
dtedit_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
server function.}
}
\description{
Shiny UI Module version of DTedit
}
\seealso{
DTedit::dtedit_server
}
