% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtedit_demo.R
\name{dteditmod_demo}
\alias{dteditmod_demo}
\title{Run a shiny app showing how the DTedit function works.}
\usage{
dteditmod_demo(...)
}
\arguments{
\item{...}{arguments pass to \code{runApp}

For example, can launch in showcase mode:

\code{DTedit::dteditmod_demo(display.mode = "showcase")}

Note that 'showcase' mode shows all
the .R files in the 'shiny_demo' directory, not
just the .R file used for this demonstration!}
}
\description{
modularized version
}
\details{
selectInputReactive
}
