% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtedit.R
\name{dteditUI}
\alias{dteditUI}
\title{Create a DataTable with Add, Edit and Delete buttons.}
\usage{
dteditUI(id)
}
\arguments{
\item{id}{the namespace of the module}
}
\description{
dteditUI - user-interface function
}
\details{
Use in conjunction with \code{callModule} and \code{dtedit} to create
editable datatables. \code{dteditUI} is used in the 'user interface' component
of the shiny app.
}
\examples{
# minimal DTedit example
\donttest{
library(shiny)
library(DTedit)

server <- function(input, output) {
  
  Grocery_List <- callModule(dtedit, 'Grocery_List',
                             thedataframe = data.frame(Buy = c('Tea', 'Biscuits', 'Apples'),
                                                       Quantity = c(7, 2, 5),
                                                       stringsAsFactors = FALSE)
                             )
}

ui <- fluidPage(
  h3('Grocery List'),
  dteditUI('Grocery_List')
)

shinyApp(ui = ui, server = server)
}
}
\seealso{
\code{\link{dtedit}} : the companion server-component function.\cr
 \itemize{
 \item \code{example("dtedit")} for a simple example.
 \item \code{dtedit_demo()} for a more complex example. Includes database interaction
 and interactions between the data of multiple datatables.
 }

Other Datatable Edit functions: \code{\link{dtedit}}
}
\concept{Datatable Edit functions}
