% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{format_to_dtedit}
\alias{format_to_dtedit}
\title{定义格式化函数为DTedit}
\usage{
format_to_dtedit(dbType)
}
\arguments{
\item{dbType}{数据类型}
}
\value{
返回值
}
\description{
定义格式化函数为DTedit
}
\examples{
format_to_dtedit()
}
