% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{format_to_sqlInsert}
\alias{format_to_sqlInsert}
\title{处理格式化函数}
\usage{
format_to_sqlInsert(dbType)
}
\arguments{
\item{dbType}{类型}
}
\value{
返回值
}
\description{
处理格式化函数
}
\examples{
format_to_sqlInsert()
}
