% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{format_to_sqlUpdate}
\alias{format_to_sqlUpdate}
\title{创建更新格式化函数}
\usage{
format_to_sqlUpdate(dbType)
}
\arguments{
\item{dbType}{数据类型}
}
\value{
返回值
}
\description{
创建更新格式化函数
}
\examples{
format_to_sqlUpdate()
}
