% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getFields}
\alias{getFields}
\title{针对控件元素进行处理}
\usage{
getFields(typeName, values, name, inputTypes, input.choices, edit.cols,
  edit.label.cols, text.width, textarea.width, textarea.height, date.width,
  numeric.width, select.width, selectInputMultiple, result)
}
\arguments{
\item{typeName}{类型名}

\item{values}{类型值}

\item{name}{名称}

\item{inputTypes}{输入类型}

\item{input.choices}{输入选项}

\item{edit.cols}{编辑字段名}

\item{edit.label.cols}{编辑字段标题}

\item{text.width}{文件宽度}

\item{textarea.width}{文件区域宽度}

\item{textarea.height}{文本区域调试}

\item{date.width}{日期宽度}

\item{numeric.width}{数值宽度}

\item{select.width}{选择赛诺}

\item{selectInputMultiple}{传递自定义函数}

\item{result}{添加数据缓存对象}
}
\value{
返回字段列表
}
\description{
针对控件元素进行处理
}
\examples{
getFields()
}
