% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtedit.R
\name{running}
\alias{running}
\title{Return if \code{uiOutput(name)} is running}
\usage{
running(name = NULL)
}
\arguments{
\item{name}{the name of the UI output.}
}
\value{
version
}
\description{
Return if \code{uiOutput(name)} is running
}
