% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{setDTtoCn}
\alias{setDTtoCn}
\title{设置DT为中文
https://datatables.net/reference/option/language}
\usage{
setDTtoCn()
}
\value{
返回值
}
\description{
设置DT为中文
https://datatables.net/reference/option/language
}
\examples{
setDTtoCn();
}
