% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtedit_test.R
\name{testDTedit}
\alias{testDTedit}
\title{test application}
\usage{
testDTedit(appname = "simple")
}
\arguments{
\item{appname}{choose test
simple
simple_modular
reactive
callback
error_test
selectInputreactive
password}
}
\value{
a shiny app
}
\description{
for testthat/codecov
}
