% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtedit.R
\name{updateDados}
\alias{updateDados}
\title{Change data in a dtedit2 table, but don´t update database
It is like a filter.}
\usage{
updateDados(dados = NULL, ui_name = NULL)
}
\arguments{
\item{dados}{New data to show - New data must have same fields.}

\item{ui_name}{dtedit2 ui_name to change}
}
\description{
Change data in a dtedit2 table, but don´t update database
It is like a filter.
}
