% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{updateData}
\alias{updateData}
\title{更新数据}
\usage{
updateData(proxy, data, ...)
}
\arguments{
\item{proxy}{代理对象，支持更新}

\item{data}{数据}

\item{...}{其他选英}
}
\value{
返回值
}
\description{
更新数据
}
\examples{
updateData()
}
