% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTedit-package.R
\docType{package}
\name{DTedit-package}
\alias{DTedit}
\alias{DTedit-package}
\title{DTedit: Editable DataTables for shiny apps}
\description{
This package extends the DT::DataTable to allow users to create, edit, and delete
    rows from the data table.
}
\author{
\strong{Maintainer}: Jason Bryer \email{jason@bryer.org}

}
\keyword{internal}
