% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTedit-package.R
\docType{package}
\name{DTedit-package}
\alias{DTedit-package}
\title{Editable DataTables for shiny apps}
\description{
Editable DataTables for shiny apps
}
\author{
\email{jason@bryer.org}
}
\keyword{DataTable}
\keyword{package}
\keyword{shiny}
