% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtedit_demo.R
\name{dtedit_selectInputReactive_demo}
\alias{dtedit_selectInputReactive_demo}
\title{Run a shiny app showing how the DTedit function works.}
\usage{
dtedit_selectInputReactive_demo(...)
}
\arguments{
\item{...}{arguments pass to \code{runApp}

 For example, can launch in showcase mode:

 \code{DTedit::dtedit_selectInputReactive_demo(display.mode = "showcase")}

 Note that 'showcase' mode shows all
 the .R files in the 'shiny_demo' directory, not
 just the .R file used for this demonstration!}
}
\description{
demonstrates interaction between datatables using
 selectInputReactive and selectInputMultipleReactive
}
