% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtedit_test.R
\name{dtedit_test}
\alias{dtedit_test}
\title{test application}
\usage{
dtedit_test(appname = "simple", ...)
}
\arguments{
\item{appname}{choose test

\itemize{
  \item \code{simple}
  \item \code{simple_modular}
  \item \code{reactive}
  \item \code{callback}
  \item \code{error_test}
  \item \code{selectInputReactive}
  \item \code{password}
}}

\item{...}{extra options passed to shiny::shinyApp}
}
\value{
a shiny app
}
\description{
for testthat/codecov
}
