% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtedit.R
\name{dteditUI}
\alias{dteditUI}
\title{Create a DataTable with external update using reactive data.}
\usage{
dteditUI(id)
}
\arguments{
\item{id}{the namespace of the module}
}
\description{
dteditUI - user-interface function
}
\details{
Use in conjunction with \code{callModule} and \code{dtedit} to create
editable datatables. \code{dteditUI} is used in the 'user interface' component
of the shiny app.
}
\seealso{
\code{\link{dtedit}} : the companion server-component function.\cr
 \itemize{
 \item \code{example("dtedit")} for a simple example.
 \item \code{example("dteditUI")} for a simple example with reactive dataframe
 \item \code{dtedit_demo()} for a more complex example. Includes database interaction
 and interactions between the data of multiple datatables.
 }

Other Datatable Edit functions: \code{\link{dtedit2}}
}
\concept{Datatable Edit functions}
